port(llvm18): SeaLoops to LLVM 18 (loops stage)
IndVarSimplify/LoopRotation carry the upstream LLVM 17->18 churn (clean 3-way
merge onto the seahorn fork); LoopUnroll needs no change. LoopUnrollPass ported
to LLVM 18's UnrollCostEstimator API: ApproximateLoopSize was folded into the
UnrollCostEstimator ctor and computeUnrollCount now takes the UCE rather than a
LoopSize; canUnroll() subsumes the duplicatable/valid-cost checks (the private
NotDuplicatable is no longer separately observable).

CMakeLists find_package LLVM 17->18.1 (LLVM 18 ships only as 18.1.x, so the
LLVMConfigVersion check requires the minor).
